refactor: conditionをmanager配下に
